Drone AI changes, HP is now double rather than uint, fix player shooting off, serialize light enabled state
Former-commit-id: 574140e586553caf549a9bd60061c0faa61e0d1c

+++ LonelyGalaxy/DamagingProjectile.hpp
@@ -3,7 +3,7 @@
 #include <stdint.h>
 namespace lg {
     struct DamagingProjectile {
-        uint64_t damage = 5;
+        double damage = 5.0;
         double creationTime = 0.0;
     };
 }
